backend/requirements.txt
Django
django-cors-headers
django-rest-framework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
python-dotenv